+++ lib/utils/app_version.dart
@@ -4,8 +4,8 @@ import 'package:path/path.dart' as p;
 import 'package:rekeens_flutter_cli/utils/project_paths.dart';
 import 'package:yaml/yaml.dart';
 
-String getAppVersion() {
-  final root = getPackageRoot();
+Future<String> getAppVersion() async {
+  final root = await getPackageRoot();
   final pubspecFile = File(p.join(root, 'pubspec.yaml'));
   if (!pubspecFile.existsSync()) return 'unknown';